feat: expose MinimapRenderer through plugin_base
Minimap.h depends on gwtoolboxdll_export.h which is
cmake-generated and only exists in the build directory.
Out-of-tree plugins couldn't use MinimapRenderer without
building GWToolboxdll first (which pulls many dependencies
and SDKs, significantly slowing down build time for
out-of-tree plugins).

- Add GWToolboxdll/RectF.h: RectF extracted from
  GuiUtils.h into a standalone header with no heavy
  dependencies
- Add GWToolboxdll/MinimapPlugin.h: RectF,
  MinimapRenderContext and MinimapRenderer with a
  self-contained export macro; RegisterRenderer/
  UnregisterRenderer are dllexported statics on
  MinimapRenderer, defined in Minimap.cpp.
  Minimap::RegisterRenderer/UnregisterRenderer are now
  private, accessible only via MinimapRenderer.
- plugin_base: expose RectF.h and MinimapPlugin.h
  via target_sources
